Conflicts:
	12_ProjetECommerce/WebContent/WEB-INF/applicationContext.xml
	12_ProjetECommerce/WebContent/index.jsp
	12_ProjetECommerce/build/classes/fr/adaming/controller/InternauteBoutiqueController.class
	12_ProjetECommerce/build/classes/fr/adaming/entity/Produit.class
	12_ProjetECommerce/build/classes/fr/adaming/test/AppTestCategorie.class
	12_ProjetECommerce/build/classes/fr/adaming/test/AppTestProduit.class
	12_ProjetECommerce/src/fr/adaming/controller/InternauteBoutiqueController.java
	12_ProjetECommerce/src/fr/adaming/test/AppTestCategorie.java
	12_ProjetECommerce/src/fr/adaming/test/AppTestProduit.java

+++ 12_ProjetECommerce/src/fr/adaming/controller/AdminProduitsController.java
@@ -155,7 +155,7 @@ public class AdminProduitsController {
 			//1. recup du fonctionnaire à modifier via son id
 		Map<String, Object> data = new HashMap<>();
 		
-		Produit produitUp = iapm.findProduitByIdService((long) pIdProduit);
+		Produit produitUp = iapm.findProduitByIdService(pIdProduit);
 		
 		List<Categorie> listeCategories = iapm.getAllCategorie();
